requirements.txt
twelvedata>=1.2.0
pandas>=2.2
pyarrow>=15.0
numpy>=1.26
pyyaml>=6.0
pydantic>=2.6
datasets>=2.18
huggingface_hub>=0.23
tqdm>=4.66
python-dateutil>=2.9
# optional: charts
plotly>=5.20
kaleido>=0.2.1
# testing
pytest>=8.0
pytest-cov>=5.0
